gpu: Renamings, cleanups
No logic change
