+++ BUILD.gn
@@ -33,7 +33,7 @@ action("vulkan_clean_old_validation_layer_objects") {
   script = "scripts/gn/remove_files.py"
 
   # inputs is a (random) new file since the vvl roll, used to ensure the cleanup is done only once
-  inputs = [ "layers/gpu_validation/gpu_validation.cpp" ]
+  inputs = [ "layers/gpu/gpu_validation.cpp" ]
   outputs = [ "$vulkan_gen_dir/old_vvl_files_are_removed" ]
   args = [
     "$raw_vulkan_gen_dir/old_vvl_files_are_removed",
@@ -166,46 +166,50 @@ vvl_sources = [
   "layers/gpu_shaders/gpu_error_codes.h",
   "layers/gpu_shaders/gpu_error_header.h",
   "layers/gpu_shaders/gpu_shaders_constants.h",
-  "layers/gpu_validation/debug_printf.cpp",
-  "layers/gpu_validation/debug_printf.h",
-  "layers/gpu_validation/gpu_constants.h",
-  "layers/gpu_validation/gpu_descriptor_set.cpp",
-  "layers/gpu_validation/gpu_descriptor_set.h",
-  "layers/gpu_validation/gpu_error_message.cpp",
-  "layers/gpu_validation/gpu_error_message.h",
-  "layers/gpu_validation/gpu_image_layout.cpp",
-  "layers/gpu_validation/gpu_settings.h",
-  "layers/gpu_validation/gpu_shader_instrumentor.cpp",
-  "layers/gpu_validation/gpu_shader_instrumentor.h",
-  "layers/gpu_validation/gpu_state_tracker.cpp",
-  "layers/gpu_validation/gpu_state_tracker.h",
-  "layers/gpu_validation/gpu_subclasses.cpp",
-  "layers/gpu_validation/gpu_subclasses.h",
-  "layers/gpu_validation/gpu_record.cpp",
-  "layers/gpu_validation/gpu_resources.h",
-  "layers/gpu_validation/gpu_settings.h",
-  "layers/gpu_validation/gpu_setup.cpp",
-  "layers/gpu_validation/gpu_validation.cpp",
-  "layers/gpu_validation/gpu_validation.h",
-  "layers/gpu_validation/gpu_vuids.cpp",
-  "layers/gpu_validation/gpu_vuids.h",
-  "layers/gpu_validation/spirv/function_basic_block.cpp",
-  "layers/gpu_validation/spirv/function_basic_block.h",
-  "layers/gpu_validation/spirv/instruction.cpp",
-  "layers/gpu_validation/spirv/instruction.h",
-  "layers/gpu_validation/spirv/link.h",
-  "layers/gpu_validation/spirv/module.cpp",
-  "layers/gpu_validation/spirv/module.h",
-  "layers/gpu_validation/spirv/pass.cpp",
-  "layers/gpu_validation/spirv/pass.h",
-  "layers/gpu_validation/spirv/type_manager.cpp",
-  "layers/gpu_validation/spirv/type_manager.h",
-  "layers/gpu_validation/spirv/bindless_descriptor_pass.h",
-  "layers/gpu_validation/spirv/bindless_descriptor_pass.cpp",
-  "layers/gpu_validation/spirv/buffer_device_address_pass.h",
-  "layers/gpu_validation/spirv/buffer_device_address_pass.cpp",
-  "layers/gpu_validation/spirv/ray_query_pass.h",
-  "layers/gpu_validation/spirv/ray_query_pass.cpp",
+  "layers/gpu/debug_printf.cpp",
+  "layers/gpu/debug_printf.h",
+  "layers/gpu/gpu_constants.h",
+  "layers/gpu/gpu_descriptor_set.cpp",
+  "layers/gpu/gpu_descriptor_set.h",
+  "layers/gpu/gpu_error_message.cpp",
+  "layers/gpu/gpu_error_message.h",
+  "layers/gpu/gpu_image_layout.cpp",
+  "layers/gpu/gpu_settings.h",
+  "layers/gpu/gpu_shader_instrumentor.cpp",
+  "layers/gpu/gpu_shader_instrumentor.h",
+  "layers/gpu/gpu_state_tracker.cpp",
+  "layers/gpu/gpu_state_tracker.h",
+  "layers/gpu/gpu_subclasses.cpp",
+  "layers/gpu/gpu_subclasses.h",
+  "layers/gpu/gpu_record.cpp",
+  "layers/gpu/gpu_resources.h",
+  "layers/gpu/gpu_settings.h",
+  "layers/gpu/gpu_setup.cpp",
+  "layers/gpu/gpu_validation.cpp",
+  "layers/gpu/cmd/draw_validation.cpp",
+  "layers/gpu/cmd/dispatch_validation.cpp",
+  "layers/gpu/cmd/trace_rays_validation.cpp",
+  "layers/gpu/cmd/copy_validation.cpp",
+  "layers/gpu/gpu_validation.h",
+  "layers/gpu/gpu_vuids.cpp",
+  "layers/gpu/gpu_vuids.h",
+  "layers/gpu/spirv/function_basic_block.cpp",
+  "layers/gpu/spirv/function_basic_block.h",
+  "layers/gpu/spirv/instruction.cpp",
+  "layers/gpu/spirv/instruction.h",
+  "layers/gpu/spirv/link.h",
+  "layers/gpu/spirv/module.cpp",
+  "layers/gpu/spirv/module.h",
+  "layers/gpu/spirv/pass.cpp",
+  "layers/gpu/spirv/pass.h",
+  "layers/gpu/spirv/type_manager.cpp",
+  "layers/gpu/spirv/type_manager.h",
+  "layers/gpu/spirv/bindless_descriptor_pass.h",
+  "layers/gpu/spirv/bindless_descriptor_pass.cpp",
+  "layers/gpu/spirv/buffer_device_address_pass.h",
+  "layers/gpu/spirv/buffer_device_address_pass.cpp",
+  "layers/gpu/spirv/ray_query_pass.h",
+  "layers/gpu/spirv/ray_query_pass.cpp",
   "layers/layer_options.cpp",
   "layers/layer_options.h",
   "layers/object_tracker/object_lifetime_validation.h",
