+++ layers/gpu/gpu_record.cpp
@@ -23,8 +23,8 @@
 #include "utils/cast_utils.h"
 #include "utils/shader_utils.h"
 #include "utils/hash_util.h"
-#include "gpu_validation/gpu_validation.h"
-#include "gpu_validation/gpu_subclasses.h"
+#include "gpu/gpu_validation.h"
+#include "gpu/gpu_subclasses.h"
 #include "spirv-tools/instrument.hpp"
 #include "spirv-tools/linker.hpp"
 #include "generated/layer_chassis_dispatch.h"
