+++ layers/gpu/gpu_state_tracker.cpp
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-#include "gpu_validation/gpu_shader_instrumentor.h"
-#include "gpu_validation/gpu_state_tracker.h"
+#include "gpu/gpu_shader_instrumentor.h"
+#include "gpu/gpu_state_tracker.h"
 
 namespace gpu_tracker {
