+++ layers/gpu/gpu_vuids.h
@@ -1,6 +1,6 @@
-/* Copyright (c) 2021-2023 The Khronos Group Inc.
- * Copyright (c) 2021-2023 Valve Corporation
- * Copyright (c) 2021-2023 LunarG, Inc.
+/* Copyright (c) 2021-2024 The Khronos Group Inc.
+ * Copyright (c) 2021-2024 Valve Corporation
+ * Copyright (c) 2021-2024 LunarG, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "gpu_validation/gpu_validation.h"
+#include "gpu/gpu_validation.h"
 
 #pragma once
 // Getter function to provide kVUIDUndefined in case an invalid function is passed in
