+++ layers/gpu/spirv/ray_query_pass.cpp
